refactor: extraer form de contrato a un archivo
