+++ src/components/design-project/DesignProjectTable.tsx
@@ -1,18 +1,12 @@
 "use client";
 
-import { useDesignProject } from "@/hooks/use-design-project";
-import { useProfile } from "@/hooks/use-profile";
 import {
     DesignProjectSummaryData,
     DesignProjectStatus,
 } from "@/types/designProject";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { ColumnDef } from "@tanstack/react-table";
-import { parse, format } from "date-fns";
-import { Contact, Download, Ellipsis, FileDown } from "lucide-react";
+import { Contact, Ellipsis, FileDown } from "lucide-react";
 import { useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
 
 import { DataTable } from "@/components/data-table/DataTable";
 import { Badge } from "@/components/ui/badge";
@@ -27,25 +21,9 @@ import {
     DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 
-import DatePicker from "../ui/date-time-picker";
-import {
-    Dialog,
-    DialogClose,
-    DialogContent,
-    DialogDescription,
-    DialogHeader,
-    DialogTitle,
-} from "../ui/dialog";
-import {
-    Form,
-    FormControl,
-    FormField,
-    FormItem,
-    FormLabel,
-    FormMessage,
-} from "../ui/form";
 import { CreateProjectDialog } from "./CreateDesignProjectDialog";
 import { DesignProjectDescriptionDialog } from "./DesignProjectDescriptionDialog";
+import { GenerateContractForm } from "./GenerateContractForm";
 
 export function DesignProjectTable({
     data,
@@ -224,7 +202,7 @@ export function DesignProjectTable({
                                 onOpenChange={setShowDescriptionDialog}
                                 project={row?.original}
                             />
-                            <ContractGenerateForm
+                            <GenerateContractForm
                                 id={row.original.id}
                                 publicCode={row.original.code}
                                 open={showContractForm}
@@ -282,123 +260,10 @@ export function DesignProjectTable({
     );
 }
 
-const contractSchema = z.object({
-    contractDate: z.string().min(1, {
-        message: "Debes seleccionar una fecha",
-    }),
-});
-
-function ContractGenerateForm(props: {
-    open: boolean;
-    onOpenChange: (open: boolean) => void;
-    id: string;
-    publicCode: string;
-}) {
-    const form = useForm<z.infer<typeof contractSchema>>({
-        resolver: zodResolver(contractSchema),
-        defaultValues: {
-            contractDate: "",
-        },
-    });
-    const { generateContractPdf } = useDesignProject();
-
-    function onSubmit(values: z.infer<typeof contractSchema>) {
-        generateContractPdf(props.id, props.publicCode, values.contractDate);
-    }
-
-    return (
-        <Dialog open={props.open} onOpenChange={props.onOpenChange}>
-            <DialogContent>
-                <DialogHeader>
-                    <DialogTitle>Generar contrato</DialogTitle>
-                    <DialogDescription>
-                        Ingresa la fecha en la que se firmará el contrato:
-                    </DialogDescription>
-                    <Form {...form}>
-                        <form
-                            onSubmit={form.handleSubmit(onSubmit)}
-                            className="space-y-8"
-                        >
-                            <FormField
-                                control={form.control}
-                                name="contractDate"
-                                render={({ field }) => (
-                                    <FormItem>
-                                        <FormLabel htmlFor="startDate">
-                                            Fecha de inicio de proyecto
-                                        </FormLabel>
-                                        <FormControl>
-                                            <DatePicker
-                                                value={
-                                                    field.value
-                                                        ? parse(
-                                                              field.value,
-                                                              "yyyy-MM-dd",
-                                                              new Date(),
-                                                          )
-                                                        : undefined
-                                                }
-                                                onChange={(date) => {
-                                                    if (date) {
-                                                        const formattedDate =
-                                                            format(
-                                                                date,
-                                                                "yyyy-MM-dd",
-                                                            );
-                                                        field.onChange(
-                                                            formattedDate,
-                                                        );
-                                                    } else {
-                                                        field.onChange("");
-                                                    }
-                                                }}
-                                            />
-                                        </FormControl>
-                                        <FormMessage />
-                                    </FormItem>
-                                )}
-                            />
-
-                            <DialogClose asChild>
-                                <Button
-                                    disabled={!form.formState.isDirty}
-                                    type="submit"
-                                >
-                                    Generar contrato
-                                </Button>
-                            </DialogClose>
-                        </form>
-                    </Form>
-                </DialogHeader>
-            </DialogContent>
-        </Dialog>
-    );
-}
-
 function DesignProjectTableToolbarActions() {
-    // TODO: bring actual data
-    //const table: Table<DesignProjectStatus> = undefined;
-    const { user } = useProfile();
-    const exportFile = false;
-
     return (
         <div className="flex w-fit flex-wrap items-center gap-2">
             <CreateProjectDialog />
-            {exportFile ||
-                (user?.isSuperAdmin && (
-                    <Button
-                        variant="outline"
-                        size="sm"
-                        onClick={() => {
-                            //if (table) {
-                            //    console.log("export on click :D");
-                            //}
-                        }}
-                    >
-                        <Download className="mr-2 size-4" aria-hidden="true" />
-                        Exportar
-                    </Button>
-                ))}
         </div>
     );
 }
